+++ src/ru/majestic/thetown/game/cargo/ICargo.java
@@ -1,5 +1,7 @@
 package ru.majestic.thetown.game.cargo;
 
+import java.math.BigInteger;
+
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
 
@@ -8,9 +10,9 @@ public interface ICargo {
    public void       load              (SharedPreferences prefs);
    public void       save              (Editor prefsEditor);
    
-   public long       getCurrentCount   ();
-   public void       add               (long count);
-   public void       remove            (long count);
+   public BigInteger getCurrentCount   ();
+   public void       add               (BigInteger count);
+   public void       remove            (BigInteger count);
    public void       clear             ();
    
 }
