+++ src/ru/majestic/thetown/game/clickers/IClicker.java
@@ -1,5 +1,7 @@
 package ru.majestic.thetown.game.clickers;
 
+import java.math.BigInteger;
+
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
 
@@ -8,10 +10,10 @@ public interface IClicker {
    public void    save                    (Editor prefsEditor);
    public void    load                    (SharedPreferences prefs);
    
-   public int     getCurrentLvl           ();
-   public int     getResourcesPerClick    ();
-   public int     getUpgradePrice         ();
-   public void    upgrade                 ();
+   public int            getCurrentLvl           ();
+   public BigInteger     getResourcesPerClick    ();
+   public BigInteger     getUpgradePrice         ();
+   public void           upgrade                 ();
    
    public String  getTitle                ();
 }
