+++ src/ru/majestic/thetown/game/market/impl/FoodMarketItem.java
@@ -1,10 +1,12 @@
 package ru.majestic.thetown.game.market.impl;
 
+import java.math.BigInteger;
+
 import ru.majestic.thetown.game.market.AMarketItem;
 
 public class FoodMarketItem extends AMarketItem {
 
-   public FoodMarketItem(long goldPrice, long productCount) {
+   public FoodMarketItem(BigInteger goldPrice, BigInteger productCount) {
       super(goldPrice, productCount);
    }
