+++ src/ru/majestic/thetown/game/market/impl/WoodMarketItem.java
@@ -1,10 +1,12 @@
 package ru.majestic.thetown.game.market.impl;
 
+import java.math.BigInteger;
+
 import ru.majestic.thetown.game.market.AMarketItem;
 
 public class WoodMarketItem extends AMarketItem {
 
-   public WoodMarketItem(long goldPrice, long productCount) {
+   public WoodMarketItem(BigInteger goldPrice, BigInteger productCount) {
       super(goldPrice, productCount);
    }
