+++ src/ru/majestic/thetown/game/town/ITown.java
@@ -1,5 +1,7 @@
 package ru.majestic.thetown.game.town;
 
+import java.math.BigInteger;
+
 import ru.majestic.thetown.game.town.listeners.OnTownNewLevelObtainedListener;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
@@ -10,9 +12,9 @@ public interface ITown {
    public void       save         (Editor prefsEditor);
    
    public int              getLvl                              ();
-   public int              getCurrentExp                       ();
-   public int              getExpToNextLvl                     ();
-   public void             addExp                              (int expCount);   
+   public BigInteger       getCurrentExp                       ();
+   public BigInteger       getExpToNextLvl                     ();
+   public void             addExp                              (BigInteger expCount);   
    
    public void             addOnTownNewLevelObtainedListener      (OnTownNewLevelObtainedListener onTownNewLevelObtainedListener);
    public void             removeOnTownNewLevelObtainedListener   (OnTownNewLevelObtainedListener onTownNewLevelObtainedListener);
