+++ src/ru/majestic/thetown/game/town/ITownLevelReward.java
@@ -1,9 +1,11 @@
 package ru.majestic.thetown.game.town;
 
+import java.math.BigInteger;
+
 public interface ITownLevelReward {
 
-   public long getFoodReward();
-   public long getWoodReward();   
-   public long getGoldReward();
+   public BigInteger getFoodReward();
+   public BigInteger getWoodReward();   
+   public BigInteger getGoldReward();
    
 }
