+++ src/ru/majestic/thetown/game/workers/listeners/OnWokersProductionCompleteListener.java
@@ -1,7 +1,9 @@
 package ru.majestic.thetown.game.workers.listeners;
 
+import java.math.BigInteger;
+
 public interface OnWokersProductionCompleteListener {
 
-   public void onWorkersProductionComplete(int addFood, int addWood);
+   public void onWorkersProductionComplete(BigInteger addFood, BigInteger addWood);
    
 }
