+++ src/ru/majestic/thetown/view/attack/IAttackView.java
@@ -1,5 +1,7 @@
 package ru.majestic.thetown.view.attack;
 
+import java.math.BigInteger;
+
 import ru.majestic.thetown.view.IClickableView;
 import ru.majestic.thetown.view.attack.listeners.OnAttackDialogClosedListener;
 
@@ -7,7 +9,7 @@ public interface IAttackView extends IClickableView {
 
    public void setOnAttackDialogClosedListener(OnAttackDialogClosedListener onAttackDialogClosedListener);
    
-   public void show           (boolean attackResult, long wood, long food);
+   public void show           (boolean attackResult, BigInteger wood, BigInteger food);
    public void close          ();
    public boolean isVisible   ();
