+++ src/ru/majestic/thetown/view/counters/ACountViewSkeleton.java
@@ -1,5 +1,7 @@
 package ru.majestic.thetown.view.counters;
 
+import java.math.BigInteger;
+
 import org.andengine.entity.Entity;
 import org.andengine.entity.primitive.Rectangle;
 import org.andengine.entity.sprite.Sprite;
@@ -38,7 +40,7 @@ public abstract class ACountViewSkeleton extends Rectangle implements ICountView
 
 
    @Override
-   public void changeCount(long newValue) {
+   public void changeCount(BigInteger newValue) {
       countText.setText(BigValueFormatter.format(newValue));
       
    }
