+++ src/ru/majestic/thetown/view/counters/ICountView.java
@@ -1,9 +1,11 @@
 package ru.majestic.thetown.view.counters;
 
+import java.math.BigInteger;
+
 import ru.majestic.thetown.view.IView;
 
 public interface ICountView extends IView {
 
-   public void changeCount(long newValue);   
+   public void changeCount(BigInteger newValue);   
    
 }
