+++ src/ru/majestic/thetown/view/counters/ICountWithMaxValueView.java
@@ -1,7 +1,9 @@
 package ru.majestic.thetown.view.counters;
 
+import java.math.BigInteger;
+
 public interface ICountWithMaxValueView extends ICountView {
    
-   public void onMaxValueChanged(long maxValue);
+   public void onMaxValueChanged(BigInteger maxValue);
    
 }
