+++ src/ru/majestic/thetown/view/counters/ICountWithRpSAndMaxValueView.java
@@ -1,7 +1,9 @@
 package ru.majestic.thetown.view.counters;
 
+import java.math.BigInteger;
+
 public interface ICountWithRpSAndMaxValueView extends ICountWithRpSView {
    
-   public void onMaxValueChanged(long maxValue);
+   public void onMaxValueChanged(BigInteger maxValue);
 
 }
