+++ src/ru/majestic/thetown/view/dialogs/shops/panels/buttons/IBuyGoldButton.java
@@ -1,5 +1,7 @@
 package ru.majestic.thetown.view.dialogs.shops.panels.buttons;
 
+import java.math.BigInteger;
+
 import ru.majestic.thetown.view.IClickableView;
 import ru.majestic.thetown.view.dialogs.shops.panels.buttons.listeners.OnBuyGoldButtonClickedListener;
 
@@ -7,6 +9,6 @@ public interface IBuyGoldButton extends IClickableView {
 
    public void    setOnBuyGoldButtonClickedListener      (OnBuyGoldButtonClickedListener onBuyGoldButtonClickedListener);
    public int     getIndex                               ();
-   public void    setGoldCount                           (int goldCount);
+   public void    setGoldCount                           (BigInteger goldCount);
    public void    setDollarsCount                        (float dollarsCount);
 }
